LRQA-41190 Move 'getSetFromString' into BaseJob so it can be used in different Jobs
