+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseJob.java
@@ -14,7 +14,12 @@
 
 package com.liferay.jenkins.results.parser;
 
+import java.util.Collections;
 import java.util.Properties;
+import java.util.Set;
+import java.util.TreeSet;
+
+import org.apache.commons.lang.StringUtils;
 
 /**
  * @author Michael Hashimoto
@@ -33,6 +38,24 @@ public abstract class BaseJob implements Job {
 
 	protected BaseJob(String jobName) {
 		this.jobName = jobName;
+	}
+
+	protected Set<String> getSetFromString(String string) {
+		if (string == null) {
+			return Collections.emptySet();
+		}
+
+		Set<String> set = new TreeSet<>();
+
+		for (String item : StringUtils.split(string, ",")) {
+			if (item.startsWith("#")) {
+				continue;
+			}
+
+			set.add(item.trim());
+		}
+
+		return set;
 	}
 
 	protected String jobName;
